requirements.txt
psycopg2-binary
requests
python-dotenv
pandas
sqlalchemy
streamlit
plotly
